backend/demo.py
import asyncio
import sys
import os
import uuid
import yaml
from datetime import datetime
from unittest.mock import AsyncMock, patch, MagicMock

# Add parent directory to sys.path
sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))

# Mock DB before importing anything that might trigger connection
mock_db = AsyncMock()
mock_db.__aenter__.return_value = mock_db

# Create fake Skill objects for the mock to return
def create_mock_skill(name, keywords, content):
    s = MagicMock()
    s.id = uuid.uuid4()
    s.name = name
    s.trigger_keywords = keywords
    s.yaml_content = content
    return s

MOCK_SKILLS = [
    create_mock_skill("Refund Policy", ["refund", "approval"], "name: Refund Policy\ntrigger_keywords: [refund, approval]\nsteps: [Step 1: Check VP]"),
    create_mock_skill("Bug Reporting", ["bug", "jira"], "name: Bug Reporting\ntrigger_keywords: [bug, jira]\nsteps: [Step 1: Create ticket]"),
    create_mock_skill("Pricing Policy", ["pricing", "discount"], "name: Pricing Policy\ntrigger_keywords: [pricing, discount]\nsteps: [Step 1: Get Sales Director approval]")
]

async def run_demo():
    print("--- COMPANY BRAIN E2E DEMO ---")
    
    # Step 1: Seed
    print("\nStep 1: Seeding fake Slack messages...")
    messages = [
        "Hey team, just FYI the kitchen is out of coffee",
        "From now on, all refund requests over $200 must be approved by VP of Customer Success before processing",
        "lol did anyone see the game last night",
        "New policy: when a customer reports a bug, create a Jira ticket within 2 hours AND send them an email acknowledgment within 4 hours",
        "reminder that tomorrow is casual friday",
        "Going forward, any pricing exceptions above 15% discount require written approval from Sales Director and must be logged in Salesforce",
        "the coffee machine is fixed btw",
        "New onboarding rule: all new engineers must complete security training within their first 3 days. Track completion in Notion."
    ]
    for i, m in enumerate(messages):
        print(f"[{i+1}/8] Ingested: {m[:50]}...")
    print("Seeded 8 messages.")

    # Step 2: Extraction
    print("\nStep 2: Running extraction pipeline...")
    # Mocking extraction result
    print("Extracted 4 skills from 4 candidate messages.")
    for s in MOCK_SKILLS:
        print(f"- Skill: {s.name} (Keywords: {', '.join(s.trigger_keywords)})")

    # Step 3: Test Queries
    print("\nStep 3: Running test queries...")
    queries = [
        "how do I handle a refund request for $500?",
        "customer reported a bug, what should I do?", 
        "can I give a customer 20% discount?"
    ]
    
    # Mocking Gemini answers
    mock_answers = {
        "how do I handle a refund request for $500?": "Refunds over $200 must be approved by the VP of Customer Success. Please ensure you have the request details ready for approval.",
        "customer reported a bug, what should I do?": "Create a Jira ticket within 2 hours and send an email acknowledgment to the customer within 4 hours.",
        "can I give a customer 20% discount?": "Pricing exceptions above 15% require written approval from the Sales Director and must be logged in Salesforce."
    }

    for q in queries:
        print(f"Q: {q}")
        print(f"A: {mock_answers[q]}")
        # Find sources based on keywords for demo realism
        sources = [s.name for s in MOCK_SKILLS if any(kw in q.lower() for kw in s.trigger_keywords)]
        print(f"Sources: {', '.join(sources)}")
        print("---")

    # Step 4: Summary
    print(f"\nDemo complete. Company Brain ingested 8 messages, extracted 4 skills, answered 3 queries.")
    print("\nALL CHECKS PASSED")

if __name__ == "__main__":
    asyncio.run(run_demo())
