backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.35
asyncpg==0.30.0
python-dotenv==1.0.1
pydantic==2.9.2
slack-sdk==3.27.2
google-generativeai==0.7.2
pgvector==0.3.2
python-multipart==0.0.9
httpx==0.27.0
